capture: use uint16_t for max_pending_packets
Use a fixed type of max_pending_packets instead of intmax_t which can
differ based on the platform/standard library.

Should also prevent lints about possible arithmetic overflow.

+++ src/runmode-unittests.c
@@ -278,7 +278,7 @@ void RunUnittests(int list_unittests, const char *regex_arg)
         UtListTests(regex_arg);
     } else {
         /* global packet pool */
-        extern intmax_t max_pending_packets;
+        extern uint16_t max_pending_packets;
         max_pending_packets = 128;
         PacketPoolInit();
